close iframe on flotiq detach event
