+++ plugins/form-secondary-column/index.js
@@ -3,8 +3,21 @@ import {
   getCachedElement,
 } from "../../common/plugin-element-cache";
 import { getCtdSettings } from "../../common/settings-parser";
+import i18n from "../../i18n";
 import pluginInfo from "../../plugin-manifest.json";
 import { createSecondaryColumn, openedState } from "./column-element";
+
+const messageEvent = (event) => {
+  if (event.data?.message !== "Live preview time updated" || !event.data?.time)
+    return;
+  const livePreviewUpdated = document
+    .querySelector(".plugin-live-preview__secondary-column")
+    .querySelector(".plugin-live-preview__status-info");
+
+  livePreviewUpdated.textContent = i18n.t("LivePreviewUpdated", {
+    time: event.data.time,
+  });
+};
 
 export const handleSecondaryColumnAdd = (
   { contentType, contentObject, formik, rerender },
@@ -16,11 +29,37 @@ export const handleSecondaryColumnAdd = (
 
   const cacheKey = `${pluginInfo.id}-${contentType.name}-${contentObject?.id || "new"}-column`;
   let panelElement = getCachedElement(cacheKey)?.element;
+
   if (!panelElement) {
     panelElement = createSecondaryColumn(rerender);
-  }
 
-  addElementToCache(panelElement, cacheKey);
+    const removeColumn = () => {
+      openedState.isOpened = false;
+      rerender();
+    };
+
+    const resizeEvent = () => {
+      if (window.innerWidth < 1024) {
+        removeColumn();
+      }
+    };
+
+    addElementToCache(
+      panelElement,
+      cacheKey,
+      () => {
+        if (!panelElement.parentElement) return;
+        panelElement.parentElement.className = "plugin-live-preview__wrapper";
+        window.addEventListener("message", messageEvent);
+        window.addEventListener("resize", resizeEvent);
+      },
+      () => {
+        window.removeEventListener("message", messageEvent);
+        window.removeEventListener("resize", resizeEvent);
+        removeColumn();
+      },
+    );
+  }
 
   return panelElement;
 };
